Remove Parcelable requirement for UiState (#37)
Let's start without this for now and see if it comes up again as something we need outside of UI. My thinking here is that we should strongly push folks toward using the data layer as a source of truth, and only use rememberSaveable as needed.

One argument could definitely be "we're not using ViewModels though, where do we keep things on config changes?", and to that I'd argue that we should be able to derive state at any moment from the backing repository. Presenters in this case should just be state factories.

If the nice compose-first API is a factor, we could also explore offering more compose `State` APIs rather than coroutines.