Pop root automatically on empty backstacks
This was an interesting case that @keder noticed, where going from a backstack of 1 to 0 doesn't automatically pop the root callback. This makes that the default behavior, while allowing for customization of this behavior. I'm not sure what the right default should be though, or if we should even allow it.

Maybe we should always require a backstack size > 0 after initialization? Do people ever want to keep backstacks around that are empty to then (re)fill later?