Setting up Macrobenchmarking and Baseline Profiles in Circuit (#250)
This PR setups a macrobenchmarking module to measure startup time with
and without a baseline profile. Measured on a Pixel 6, API 33.

**StartupBenchmark: Without Baseline Profile**
StartupBenchmark_startup
timeToInitialDisplayMs   min 219.4,   median 224.7,   max 240.8

**StartupBenhmark: With Baseline Profile**
StartupBenchmark_startup
timeToInitialDisplayMs   min 212.7,   median 215.4,   max 241.6

We are seeing about a 4% improvement with the min and median times. Zac
mentioned that because this is a pretty basic app, the performance wins
are not that apparent. Would be more apparent in a more complex app like
Slack!

Co-authored-by: Zac Sweers <pandanomic@gmail.com>