Update NavigatorImpl.pop to invoke onBackPressedDispatcher
The current back solution seems to work when tapping the system back button or triggering the
back gesture. However, manual calls to Navigator.pop (triggered by an on screen back button)
when viewing the root backstack record will crash the app. This happens purposely as we don't
want to allow an empty backstack.

To resolve this, NavigatorImpl has been updated to invoke a onBackPressedDispatcher whenever
pop() is called while on the root backstack record.