Fix accidental minSdk 28 in circuit-retained and split baseline profiles (#449)
Resolves #259 and resolves #448 along the way as I needed to fix it to
use it in the sample.

This adds a script to split baseline profiles into their corresponding
artifacts! It uses the original fat profile and `awk`'s matching lines
into relevant other artifacts, then finally prunes the moved lines from
the original shared one.

In order to capture profiles for circuit-retained, I added a trivial
usage in the sample so its use would be captured in a profile.