Update where we setup BackHandler for Android (#470)
### Problem
#433 (part of efforts to improve KMP support) renamed
`BaseNavigableCircuitContent` to `NavigableCircuitContent`, overloading
the existing Android-specific method with the same name. The only
difference between the two method signatures was an additional param in
the android version (related to setting up the `BackHandler`).

As a result of this change existing uses of `NavigableCircuitContent` on
Android started calling the wrong method, bypassing the logic that
configured the system back handler.

### Fix 
Rather than change the Android overload in some way, or require that
Android callers provide the extra parameter (a friction when 99% of uses
will want the default), I opted to move the back handler setup logic
into the Android-specific `rememberCircuitNavigator`, and remove the
Android overload entirely.