Remember Presenter and UI in CircuitContent (#726)
Currently the `Presenter` and `Ui` are created every time
`CircuitContent` recomposes, this causes flashes of content each time it
is called.

For most use cases this won't cause an issue, since `CircuitContent`'s
inputs don't change meaning no recompositions, but it is an issue for
use-cases which need to recall `CircuitContent`.

A use-case I'm playing with at the moment is putting `CircuitContent`
within a `movableContentOf`, so `CircuitContent` will be called multiple
times. This PR changes CircuitContent to remember the Presenter and Ui,
and thus moving CircuitContent composables now works as expected.