Make CircuitContent overload with Navigator public (#727)
Not sure if there's a reason why this overload is internal, but it's
useful for custom animations and wrapping CircuitContent in
movableContentOf.

My use case is a custom `NavigableCircuitContent` which provides
animation and gesture support. The only public overload which kinda
supports this is via `onNavEvent`.
I could proxy events to the host navigator, but some of the NavEvent
subtypes are internal too.

I'm open to this change, or making the NavEvent subtypes public.