Remove intermediate event sinks
I can no longer reproduce the original issue. While the recomposition factor does come into play, this seems more or less acceptable that the recomposition can happen when the state updates. This simplifies our usage a fair bit, and I kind of wonder if the original bug was something else wrong with the compose code we'd written at the time. I think we should run ahead with this until/unless we see reproducible issues with it.