Fix EventListener.start callback order (#740)
I need to do some more testing, but I believe that `EventListener.start`
can end up getting called after `EventListener.onBeforeCreateListener`
because the body of `DisposableEffect` runs in parallel with the rest of
the composition, and therefore can race.

I can reproduce this reliably with an EventListener like the one below
```kotlin
class MyEventListener(
    private val screen: Screen,
    private val context: CircuitContext
) : EventListener {
  private lateinit var rootTransaction: ITransaction
  private lateinit var createPresenterSpan: ISpan

  override fun start() {
    rootTransaction = Sentry.startTransaction(screen.javaClass.simpleName, "circuit")
  }

  override fun onBeforeCreatePresenter(
    screen: Screen,
    navigator: Navigator,
    context: CircuitContext
  ) {
    createPresenterSpan = rootTransaction.startChild("createPresenter")
  }
  ...
}
```

Crashes because `rootTransaction` has not been initialized when
`onBeforeCreatePresenter` is called.

I think we should pull the `eventListener.start()` call out of the
`DisposableEffect` should do the trick. Maybe invoke it inside the
`remember` call when creating the `EventListener`.