Fix baseline profile generation (#768)
~Still currently no longer produces anything when I run `./gradlew
:backstack:generateBaselineProfile`, seems like something regressed with
libraries or KMP support or both~

Got this working well enough for our needs now. This doesn't generate a
startup profile for the star app currently due to a bug that was found
while diagnosing issues with this. The rest of the changes are good to
go though and we can use this while we wait for a fix in a new release.
Ref: https://issuetracker.google.com/293889189

This removes baseline profiles from the circuit-retained artifact, as
I'm currently unable to get it to pick up any on the startup hot path. I
think that's ok though 👍