JS Support (#801)
Resolves #649

This introduces JS support to the core Circuit artifacts (via
`nodejs()`) and also adds a browser app to the `counter` sample.

Some things along the way
- Adding an expect/actual indirection for the `TestEventSink` supertype
to support JS as JS compilation doesn't allow implementation of function
types.
- Consolidating around recommending `Screen` object types be `data
object` for better toString()-ability. We can't use
`Screen::class.simpleName` in common code due to JS reflection
limitations, so using this as an opportunity to lean into `data object`
more since the lack of a toString() on objects was initially why we
didn't do that.
- Extract the Counter UI in the counter sample to `common`. Will
eventually port Desktop to this too, but saving that for a later PR.
- Moved `SwiftSupport` in counter sample to iOS source set.



https://github.com/slackhq/circuit/assets/1361086/a47ef192-c70e-4958-a989-790c4010561f

---------

Co-authored-by: Jake Wharton <github@jakewharton.com>