Tweak the fix for #799
The current implementation uses the presenter itself
as the key(), which does fix the original issue. For our
usage though, the presenter will always be a new instance,
and therefore there will never be a consistent key across
back stack changes (i.e. a push, then pop). This then forces
the currentCompositeKeyHash for the underlying content
underneath to change, which then breaks rememberRetained() when
used without an explicit key.

Fixed by changing the key to the `screen` instance instead, which
achieves the same original fix, but the screen is consistent across
navigation changes, which allows the currentCompositeKeyHash to be
consistent.