Allow retained state to be retained whilst UIs and Presenters are on the back stack (#888)
This PR contains a bunch of subtle changes, to enable retained state to
be retained whilst UIs and Presenters are on the back stack. To see this
in action, you can see this [draft
PR](https://github.com/chrisbanes/tivi/pull/1551) on Tivi. The videos
attached to it highlight the benefits pretty clearly 🎉.

This follows on from a big discussion on Slack about the best way to
keep state around so that presenter cold starts don't mean that we're
showing empty states when coming back from the back stack.

## So what are the changes?

### Nested RetainedStateRegistry

Currently `RetainedStateRegistry` is designed to be used single-y,
having a global registry used at the root of the content. Since we need
to manage the lifecycle of retained state differently per back stack
entry, we need a way to be able to 'save' retained state at different
times. This PR implements this by making `RetainedStateRegistry`
nested-aware, such that registries practically become a tree structure.
In practice, this just means a small change to `RetainedStateRegistry`
so as `saveAll()` is proxied to any child registries.

### CanRetainChecker changes

We use `CanRetainChecker` as the way to control the lifecycle of
retained state. This PR changes the default checker to either
`CanRetainChecker.Always` (on non-Android platforms), or makes it nested
registry aware on Android.

### collectAsRetainedState

I found this super useful in Tivi. I initially thought about retaining
the `Presenter`s last emitted `CircuitUiState` (a bit like
`rememberUpdatedState`), but the UI state typically contains some
transient properties (`isLoading`, `errorMessage`, etc) which don't make
sense to retain. Using `rememberRetainedState` or
`collectAsRetainedState` allows retaining of the more long-lived
properties only.

### Other

- `continuityRetainedStateRegistry()`. Moved this to `commonMain` with
simple implementations for non-Android platforms.
- `SaveableHolder` -> `RetainableHolder`
- Updated Star sample to use a `continuityRetainedStateRegistry`

## Future

There's a few things that this PR doesn't tackle:

- Retaining state for different navigation roots. If the user switches
the root navigation entry (i.e via a bottom navigation bar), it would be
nice to (optionally) retain the state of each root, similar to AndroidX
Navigations `saveState` flag. Should be fairly easy, just need to update
the `CanRetainChecker` to retain `if screen was root` (or something).

---------

Co-authored-by: Zac Sweers <pandanomic@gmail.com>