Add codegen mode to support both Anvil and Hilt (#963)
This PR implements a KSP argument for `CircuitSymbolProcessor` which
allows a consumer to skip emission of Anvil annotations. One use case of
this is to allow for generated binding support for other DI frameworks,
such as Hilt.

The argument "circuit.generate-anvil-bindings" defaults to true, and can
be configured via the KSP gradle DSL:

```kotlin
ksp {
  arg("circuit.generate-anvil-bindings", "false")
}
```