Commonize SaveableStateRegistryBackStackRecordLocalProvider (#1037)
Commit 1:

Commonizes `SaveableStateRegistryBackStackRecordLocalProvider` to be
supported across all currently supported platforms by adding it to the
default `LocalBackStackRecordLocalProviders`.

The only change this requires was bringing in
[atomicfu](https://github.com/Kotlin/kotlinx-atomicfu) for a
multiplatform `synchronized` call to replace the jvm-only one.

Commit 2:

Adds tests in `circuit-foundation` to verify the new behavior on both
jvm and Android, based on the current retained tests. This required a
bit of work to allow the same test to run on both, by setting up a
delegating `ComposeUiTestRunner` and typealiased `Parcelize`.

Commit 3:

While I was there, I commonized the rest of `circuit-foundation` tests
to run across both jvm and Android. Not directly related to fixing #1036
, can remove or put into a new PR if desired.