Fix dropping back stack retained state on Android Activity rotations (#1063)
We store the RetainedStateRegistries for each back stack entry into an
'navigation content' `RetainedStateRegistry`. If we don't do this, those
registries would be stored directly in the current
`LocalRetainedStateRegistry` value, which will mostly likely be the
`continuityRetainedStateRegistry`. On Android, that
`continuityRetainedStateRegistry` will drop any 'unclaimed' values when
the host Activity is recreated. Since records on the back stack aren't
attached to composition, they can't claim their retained registries and
thus we drop all of the state for the record.

Using this 'navigation content' registry means that _it_ will be stored
in the `continuityRetainedStateRegistry` instead, and any back stack
record registries stored within the 'navigation content' registry. The
difference is that `NavigableCircuitContent` will be attached to
composition for the entire lifetime that we care about, and thus will be
able to save/claim the 'navigation content' registry on recreations.
Since any back stack registries are nested in this 'navigation content'
registry, everything is saved/claimed correctly.

Added an Android instrumentation test to stop this regressing in the
future.

Fixes #1046