Add ability to customize ModalBottomSheet appearance in BottomSheetOverlay (#1059)
I've read that you're not accepting external contributions - raised the
limitations of the `BottomSheetOverlay` in [a
discussion](https://github.com/slackhq/circuit/discussions/1048) - but
thought it might be a quick win if we could incorporate those changes
directly.

This change adds the ability to customize the `ModalBottomSheet`s
appearance:
* `dragHandle`
* `sheetShape`
* `containerColor`

Additionally it adds the ability to specify if the partiallyExpanded
state of the sheet should be skipped. (There are use cases where the
sheet should directly expand to it's full height)

Would be great if this could be incorporated.
Let me know if the optional parameters in the existing constructors are
fine or there should be dedicated ones.