Support animating an overlay out after returning a result (#1066)
Currently, if we want an exit animation for an overlay, we must wait for
that animated to complete before returning the [Result]. This results in
a lot of perceived lag if the [Result] should update the screen it was
launched from. While this can be worked around with injecting callbacks
or composition locals, those aren't ideal.

This updates `ContentWithOverlays` to animate an `AnimatedOverlay` in
and out instead of just adding and removing it from composition. I'm
open to ways to improve this.