Drop Counter iOS app deployment target to iOS 16.0 (#1159)
The `IPHONEOS_DEPLOYMENT_TARGET` is basically the minimum iOS version
the app supports. For a sample, 17.2 is probably fine, but the GHA
images tend to lag public Xcode versions so it makes sense to drop it a
little (iOS 17.2 requires Xcode 15.2).

Dropped the deployment target to 16.0, which is more realistic as a
real-world app.