Fix CI emulator tests (#1176)
- The recent change to `macos-14` means that we can't use x64 emulator
(as currently set). Ideally we would just use ARM64 emulators, but that
is blocked by
https://github.com/ReactiveCircus/android-emulator-runner/issues/350
though.

- Instead I've refactored the CI setup:
- Moved most things to run `ubuntu-latest`. We can use the new
nested-virt KVM support to make the emulator fast there.
- We still use `macos-14` but only for iOS builds in a separate job.
This is similar to what I do in Tivi.
- Add some hierarchy to the jobs, so that `publish` only happens when
all of the other jobs pass.

- Benchmark tests do not work (by default) on emulators, which is why
they've been failing for a while. We can suppress that error and get
them to run but there's little benefit, so it's better to just disable
them when running on an emulator.