Add support for inter-Screen results (#1154)
This adds support for passing results back in inter-screen nav.

For review, I'd suggest focusing on `BackStack`, `SaveableBackStack`,
`Navigator`, and `AnsweringNavigator`. There's some other file noise due
to dependency classpath updates and aligning naming `backStack`
variables/params.

## Usage

```kotlin
val filterScreenNavigator = rememberAnsweringNavigator<FilterScreen.Result>(navigator) { result ->
  filters = result.filters
}

// Elsewhere
filterScreenNavigator.goTo(FilterScreen(filters))

// In FilterScreen.kt
data class FilterScreen(...) : Screen {
  @Parcelize
  data class Result(val filters: Filters) : PopResult
}

class FilterPresenter {
  @Composable fun present(): State {
    // ...
    navigator.pop(result = FilterScreen.Result(newFilters))
  }
}
```

State restoration (config changes and process death) handled
automatically, launch ownership (i.e. only the presenter that requested
it can read the result), is owned by the backstack record and GC'd with
it.

## Under the hood

There are two core areas of implementation

### BackStack

The back stack functions as the mailbox/pipeline for passing results
along. I tinkered with a few different areas at the navigator level, but
this never felt right as the navigator is just a light shim over the
back stack. Every API iteration of this resulted in just bolting on
copies of the back stack APIs to the navigator, and eventually I felt it
was best just to manipulate the backstack directly via the
`rememberAnsweringNavigator` API instead.

Back stack `Record`s now expose an API to await a result via
`awaitResult()`, and the parent `BackStack` is responsible for giving it
information during push/pop functions.

`push()` now accepts a `resultKey: String` param, and this is a key that
must be passed to `awaitResult()` on the other side. This is to ensure
that only the caller that launched for result gets it back. When a new
screen is pushed on to the top of the stack, the back stack sets this
key on the previous record for reference.

`pop()` now accepts an optional `PopResult`, which is the marker
interface that all results must implement. Like `Screen`, it's also
parcelable on android and must be saveable since it is persisted with
the Record. This is also exposed on the `Navigator.pop()` call.

Internally in `SaveableBackStack.Record` (the only impl of this right
now), it contains a `resultChannel` with a capacity of 1 and overflow
behavior to drop latest. This ensures only one result is ever kept, and
new results replace the old (if unconsumed). When a new record is
pushed, it also clears any pending result on the previous record. This
channel more or less acts as the plumbing for the result, and the record
just does some extra checks to make sure the caller awaiting has a
matching key to receive it.

### `rememberAnsweringNavigator`

This composable function encapsulates all the logic for navigating to a
screen with the expectation of a result. This manages generation of a
key to give to the record, and manipulates the backstack directly
(similar to how `rememberCircuitNavigator` does), allowing us to keep
these the added API surface area to the lower level record APIs.

This manages tracking the back stack state and, upon returning to our
previous record, will then automatically await a result of the expected
type and then passes it on to the receiver `body` passed to this
function.

This works great with the key because upon restoration it'll just pick
up listening where it left off, and the `goTo` is done separately. This
means if it's already on the back stack behind another screen, it does
nothing and will automatically resume when it comes back to the top of
the stack. This takes some inspiration from the androidx activity result
APIs for compose too.

## More notes

Some caveats
- Result unpacking requires just type checks. Could maybe introduce
something like `ResultingScreen<T : PopResult>` to let the compiler
enforce some of this. Not sure about the screen sending the result
though. I did try the interface briefly, but feel it may be limiting in
the end because it doesn't allow for a screen to return multiple
different types.
- ~Breaks if you have an intercepting navigator that doesn't delegate to
a real one backed by a backstack. Requires normalizing passing the
navigator on as part of state, which feels odd~ Fixed ✅

> Is there a place for both this and overlays?

Yes, I believe so. Overlay is still useful for a tighter suspending +
type-safe contract, non-saveable inputs/outputs, ephemeral UIs, and the
ability to have live stable state classes.

**Next steps**

- [x] Tighten API semantics.
- [x] ~I'm not super happy with the resultKey API, and wonder if we can
further hide it. Maybe something involving exposing the backstack record
via CompositionLocal?~ Think I got a good solution for this!
- [x] More docs (namely site docs, pending code review)
- [x] More tests

**Open questions**
- [x] ~Can we hide `resultKey` from the navigator API? Maybe expose the
underlying record directly via `peek()` and bypass the navigator
intermediary.~
- Settled for exposing on the backstack but not in the navigator. Feels
like a good middle ground, as the backstack/record API is lower level
and something a navigator just utilizes under the hood.
- This revealed some issues with `FakeNavigator`, should we move it to
just `ArrayDeque` for everything it records rather than `Turbine`?
- Should we error if this is called with no backstack present, or
gracefully fall back to a navigator? My feeling is the latter, but worry
about silent failures. One tricky thing is this effectively hands
navigation access potentially to content areas that the user may not
want to offer. Maybe we need to make a `NonNavigableContent { }` block
to undo this?
- We have some inconsistencies around whether it's possible for a
backstack to be empty or not. Technically we allow it, but we almost
always treat that as an exceptional situation. Should we codify that?
- Should we make a generic `ResultingScreen<R>`? Could this help make
returning results more type-safe too?
- ~Is `GoToNavigator` the right abstraction? My thinking was it could be
nice for indirection (i.e. `val navigator = if (canDoResults)
rememberAnsweringNavigator(...) else realNavigator`).~
- Ended up being required for the new impl. Naming could maybe be
improved though.