Implement ToastEffect
A composable function that returns a lambda to show a [Toast]. Any previously shown toast will be cancelled when a new one is shown or this composable exits composition. The returned lambda can be called with the text to show in the toast.

```kotlin
val showToast = toastEffect()
// ...
Button(onClick = { showToast("Hello, world!") }) {
  Text("Show Toast")
}
```