Add a `key` to CircuitContent to keep UI and Presenter consistent  (#1254)
In order to resolve #1169 this keys the ui and presenter together so
they behave the same.

Here's the *CircuitContent cases this covers: 
#### 1. NavigableCircuitContent #### 
- The ui and presenter are retain based on the record, as the `Screen`
can't change without a new record instance.

#### 2. CircuitContent #### 

- **a)** Each `Screen` instance is a new "page" and would behave the
same as `NavigableCircuitContent`, by being keyed on the `Screen`
instance.
- **b)** The `Screen` is a model, and it's used to update the current
`CircuitContent` state, as if it were another Compose element. This is
potentially common with a "widget" sub-circuit case.