Check canRetainCheck when saving RetainedStateRegistry (#1276)
This is mostly @alexvanyo's work from #1253. Unfortunately I couldn't
find a way for a PR from a fork to be based on another PR from a
different fork.

My changes add the relevant fixes to make sure that saved/restored back
stacks (aka back stacks) work with the changes. Primarily this is so
that we check both the current back stack _and_ saved backstacks when
determining whether we can retain state.

Fixes #1252

---------

Co-authored-by: Alex Vanyo <vanyo@google.com>