Make RememberObservers in nested registries work (#1281)
At the moment `RememberObserver` support for retained state only works
if it is added directly to the root retained registry. That works fine
in our simple test, but `NavigableCircuitContent` adds a much more
complex retained registry system, which uses multiple levels of
registries.

This PR fixes this, adding support for nested registries. I've confirmed
that this works as expected in Tivi.