Deprecate assertIsEmpty/expectNoEvents, add versions for all event types (#1375)
Sometimes it's nice to be able to assert that you _haven't_ gotten a pop
event. The API previously didn't allow this. Also noticed that
assertIsEmpty and expectNoEvents were only operating on the goToEvents
turbine. I don't feel strongly about the "deprecate and add multiple
methods" approach, I'd also be fine with just keeping the top-level
methods and just adding calls to popEvents/resetRootEvents

---------

Co-authored-by: Eric Denman <1063557+edenman@users.noreply.github.com>
Co-authored-by: Zac Sweers <pandanomic@gmail.com>