Implement support for stateless UIs (#1418)
Rehash of #565. This implements support for an optimized stateless UI.
These UIs either have no state or derive their state from the screen
inputs.

There's three parts to this:
1. The introduction of internal `StaticPresenter` and `StaticUi`
classes. The former is just a presenter that always returns the same,
lazily-computed state value. The latter is a UI that always renders the
same initial static value (which can be derived from a Screen input or
otherwise). Internally in `CircuitContent`, `StaticUi` UIs have a
different code path that never wires up the presenter and instead just
renders the UI directly and only the UI.
2. The introduction of a "stateless" presenter, which is just a
singleton `StaticPresenter` instance that can be reused for circuit's
internal wiring.
3. The introduction of a `StaticScreen` interface, which is a marker
interface to indicate that a given screen is stateless and thus may not
have a presenter. `Circuit` internally checks this and will just return
the stateless presenter singleton upon these cases.

The `AboutScreen` in the STAR sample has been updated for this use case.
This is particularly useful for small isolated screens where the
`Screen` class alone can be the only input but a `CircuitContent` is
still desirable for indirection.

I left some toe-holds for possibly exposing some of these APIs down the
line, but kept things limited for now.