Add lifecycle aware Presenters (#1282)
This PR adds a new `Lifecycle` interface, which presenters and UI can
observe to know when they are 'paused'. The API is rudimentary at the
moment and will change before this is ready to land.

A bundled `PauseablePresenter` class can be extended, enabling clients
to automatically add pausing ability to their presenters. This impl will
simply return the last emitted `UiState` when the presenter is paused.
Again, name TBD.

### Other things:

- `GestureNavigationRetainedStateTest` and
`GestureNavigationSaveableStateTest` have been combined into a
parameterized `GestureNavigationStateTest`. This new test also now tests
`CupertinoGestureNavigationDecoration` so we get extra coverage.
- Changed `rememberRetained`'s `key` param to `Any` to be consistent
with all of the other `remember` functions.

### TODO:

- [x] Saveable is currently broken by these changes. This needs to be
fixed before landing.
- [x] Remove all of the logging code.
- [x] Add comments
- [x] Fix the last remaining failing test