Add withCompositionLocalProvider to avoid backward writes
This allows us to optimize pausableState, and migrate away from
a mutable state to store the last emitted state. We now simply
return the returned value, and store it in a simple mutable holder,
avoiding unnecessary recompositions due to backward writes.

Unfortunately we had to copy in `SaveableStateHolder` and tweak it
to return a value from `SaveableStateProvider`.