Commonize remaining star networking operations (#1468)
This PR converts the remaining networking logic of the star sample to
common logic using ktor. Bio parsing remains JVM only, but we can look
for something like a ksoup library down the line. Still uses OkHttp as
the engine under the hood.

One notable change is that this uses 401 responses as an indicator that
the auth token needs to refresh, as that's all ktor supports out of the
box (no expiration handling). For our cases, this is fine. In a "real"
app, we'd write a custom plugin that implements the original logic
instead.

Sadly this removes Retrofit, as it's not really valuable to use both
anymore. There is a "ktorfit" library out there, but I found it surplus
to the requirements in this case.

This also adds some neat ktor extensions for eithernet, which I'll
probably look to upstream in some way. I'm not super familiar with ktor
yet, it would be neat if it had something that allowed us to replicate
the CallAdapter behavior we used to up-convert results to ApiResult
wrappers.