Add support for kotlin-inject-anvil (#1635)
This introduces support for kotlin-inject + kotlin-inject-anvil. Note
that kotlin-inject needs another release with
https://github.com/evant/kotlin-inject/pull/414 to work, so I'm game
putting this in the release now but mentioning this caveat. Will remove
the snapshots use and just disable the new sample project for now.

kotlin-inject-anvil has support for options coming too, but in the
meantime I think it's actually fine to check in an anvil hint property
too as it doesn't require any runtime anvil APIs and can just be
compileOnly.

**TODO in the future**

- [x] Need a new kotlin-inject release for
https://github.com/evant/kotlin-inject/pull/414
- [x] Need anvil support for annotation options:
https://github.com/amzn/kotlin-inject-anvil/issues/24
- [ ] Migrate STAR sample (can be a later PR if need be)