Change SaveableStateHolder to remove saved states of removed records (#1922)
As described in #1895, `NavigableCircuitContent` internally uses
`SaveableStateHolder` to store and restore saveableState when a record
changes. However, there is an unintended issue where the state persists
even for popped records.

In this change, a `DisposableEffect` has been added after rendering the
contents of `SaveableStateHolder`. This ensures that when a specific
record is removed, removeState is called before the values are saved
within `SaveableStateHolder`, preventing unintended state retention.

Fixes #1895

---------

Co-authored-by: Zac Sweers <pandanomic@gmail.com>